+++ clients/test/unit/test_node_internals.py
@@ -1,13 +1,5 @@
 import unittest
 import yangvoodoo
-import subprocess
-
-
-command = 'sysrepocfg --import=../init-data/integrationtest.xml --format=xml --datastore=running integrationtest'
-process = subprocess.Popen(["bash"], stdin=subprocess.PIPE, stdout=subprocess.PIPE)
-(out, err) = process.communicate(command.encode('UTF-8'))
-if err:
-    raise ValueError('Unable to import data\n%s\n%s' % (out, err))
 
 
 class test_node_based_access(unittest.TestCase):
@@ -15,8 +7,8 @@ class test_node_based_access(unittest.TestCase):
     def setUp(self):
         self.maxDiff = None
         self.subject = yangvoodoo.DataAccess()
-        self.subject.connect()
-        self.root = self.subject.get_root('integrationtest')
+        self.subject.connect('integrationtest')
+        self.root = self.subject.get_root()
         self.schemactx = self.root._context.schemactx
 
     def test_get_yang_type_simple_base_case(self):